+++ source/Octopus.Client/Repositories/Async/EventRepository.cs
@@ -24,8 +24,8 @@ namespace Octopus.Client.Repositories.Async
             string eventCategories = null,
             string tenants = null,
             string tags = null,
-            int? fromAutoId = null,
-            int? toAutoId = null);
+            long? fromAutoId = null,
+            long? toAutoId = null);
     }
 
     class EventRepository : BasicRepository<EventResource>, IEventRepository
@@ -63,8 +63,8 @@ namespace Octopus.Client.Repositories.Async
             string eventCategories = null,
             string tenants = null,
             string tags = null,
-            int? fromAutoId = null,
-            int? toAutoId = null)
+            long? fromAutoId = null,
+            long? toAutoId = null)
         {
             return Client.List<EventResource>(Client.RootDocument.Link("Events"), new
             {
